refactor(views): split viewsets/handler.rs by responsibility (#4313)
Decompose the 1674-line viewsets/handler.rs into responsibility-based
submodules under viewsets/handler/ following Rust 2024 Edition module
conventions. Public API preserved via pub use re-exports.

Fixes #4313
